requirements.txt
fastapi==0.75.0
firebase-admin==5.2.0
gunicorn==20.1.0
python-dotenv==0.19.2
python-multipart==0.0.5
requests==2.27.1
uvicorn==0.17.6